refactor(browseros): move extension feeds module into release/extensions package
