+++ packages/browseros/bos_build/release/extensions/__init__.py
@@ -1,0 +1,3 @@
+#!/usr/bin/env python3
+"""Extension release package: packaging specs, CRX build/pack/upload, and
+coherent update-manifest regeneration through the feeds publisher."""
